Abstract custom item ping GUI to a generic property list modification GUI
