+++ src/main/java/tomato/Tomato.java
@@ -11,7 +11,6 @@ import tomato.backend.data.TomatoData;
 import tomato.gui.TomatoGUI;
 import tomato.gui.chat.ChatGUI;
 import tomato.gui.maingui.CustomPingGUI;
-import tomato.gui.maingui.ItemPingGUI;
 import tomato.gui.maingui.TomatoBandwidth;
 import tomato.gui.maingui.TomatoMenuBar;
 import tomato.gui.warnings.JavaOutOfMemoryGUI;
@@ -198,6 +197,7 @@ public class Tomato {
     private static void bootload(TomatoData data) {
         data.bootload();
         CustomPingGUI.loadIdPing(data);
-        ItemPingGUI.loadIdPing(data);
+        data.loadPropList("entityIdPings");
+        data.loadPropList("itemPings");
     }
 }
